fix issue #10: generalized endpoint / uri retrieval for all controllers in LdbController trait
